% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_load.R
\name{add_com_sec_names}
\alias{add_com_sec_names}
\title{Add sectors and commodities names in the simulation output database based on a names_ files}
\usage{
add_com_sec_names(
  data,
  commodities_names = names_commodities,
  sectors_names = names_sectors,
  exception_s_c = c("CONS", "CONT")
)
}
\arguments{
\item{data}{data.frame output of simulations}

\item{commodities_names}{data.frame containing codes and names of commodities}

\item{sectors_names}{data.frame containing codes and names of sectors}

\item{exception_s_c}{vector fo 4 characters string that are neither sectors or commodities}
}
\value{
the data where the names of sectors and commodities appear in the sectors and commodities database
}
\description{
Add sectors and commodities names in the simulation output database based on a names_ files
}
