% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_env.R
\name{calibration_environment}
\alias{calibration_environment}
\title{Creating the scenario calibration environment}
\usage{
calibration_environment(baseline_calibration = FALSE, configuration = config)
}
\arguments{
\item{baseline_calibration}{Boolean. Whether to create an environment to test a baseline scenario calibration. Default: FALSE}

\item{configuration}{Configuration list to use. Defaul is config. Seldom needs user modification.}
}
\value{
a list of all the objects necessary to test a scenario configuration script. the list needs to be expanded in the global environment
}
\description{
The scenario calibration environment, a.k.a. the calibration bubble, creates the necessary objects from a given configuration list to test scenario calibration scripts.
}
\examples{
\dontrun{
  config <- readconfig(
  input_config_file = file.path("configuration", "config_input_threeme.R"),
  output_config_file = file.path("configuration", "config_output_threeme.R"))

  calibration_environment(FALSE, config) |> list2env(envir = globalenv())

}


}
