% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_table.R
\name{calibration_table}
\alias{calibration_table}
\title{calibration table}
\usage{
calibration_table(params_calib, project_name, lang = NULL)
}
\arguments{
\item{params_calib}{vector of the parameters from the ThreeME model with the code of the parameters used in the model (e.g.: ES_CHD)}

\item{project_name}{name of the project in which the values should be extracted}

\item{lang}{language chosen, two options "english" and "french"}
}
\value{
a gt object
}
\description{
calibration table
}
