% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confetti.R
\name{confetti}
\alias{confetti}
\title{Create a theme for tables and plots}
\usage{
confetti(
  lines = c("#00496FFF", "#BF812DFF", "#0F85A0FF", "#7D4F73FF", "#87AFD1FF",
    "lightgoldenrod2"),
  background = "whitesmoke",
  text = "grey30"
)
}
\arguments{
\item{lines}{Vector of colours for the graph lines}

\item{background}{One colour for table background}

\item{text}{One colour for texts}
}
\value{
a list of colours, named for the tables and graphs
}
\description{
Create a theme for tables and plots
}
\examples{
palette_a <- confetti()
}
