% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_addin.R
\name{config_code}
\alias{config_code}
\title{Deparse a value for a configuration file}
\usage{
config_code(x)
}
\arguments{
\item{x}{the value.}
}
\value{
R source text.
}
\description{
Deparse a value for a configuration file
}
\keyword{internal}
