% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib_longformat.R
\name{contrib_calc}
\alias{contrib_calc}
\title{contrib_calc}
\usage{
contrib_calc(...)
}
\arguments{
\item{...}{contrib_longformat arguments, see contrib_longformat documentation}
}
\value{
same as contrib_longformat
}
\description{
Wrapper function for contrib_longformat
}
