% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib_longformat.R
\name{contrib_longformat}
\alias{contrib_longformat}
\title{Contributions calculations for long format}
\usage{
contrib_longformat(
  data,
  var1,
  var2,
  indicator = "rel.diff",
  scenar = scenarios,
  check_digit = 3,
  neg.value = NULL
)
}
\arguments{
\item{data}{data.frame where to get the data from}

\item{var1}{character string. main variable}

\item{var2}{character string vector, length >1 , variables that contribute to var1}

\item{indicator}{"rel.diff" for calculations in relative difference for two scenarios. "gr.diff"  for calculations  for one scenario , "share" , for simple share calculations}

\item{scenar}{scenarios, max 2}

\item{check_digit}{level of precision expected}

\item{neg.value}{variables that should contribute negatively (ie Imports in the case of GDP contributions calculation)}
}
\value{
a data.frame with the contributions computed
}
\description{
Contributions calculations for long format
}
