% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary_languages}
\alias{dictionary_languages}
\title{Languages the dictionary carries}
\usage{
dictionary_languages()
}
\value{
a character vector of language codes.
}
\description{
Languages the dictionary carries
}
\keyword{internal}
