% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmodelsolver.R
\name{format_duration}
\alias{format_duration}
\title{A duration, in the unit that reads best}
\usage{
format_duration(seconds)
}
\arguments{
\item{seconds}{a number of seconds.}
}
\value{
a string: \code{"0.52 s"}, \code{"38.4 s"}, \code{"1 min 14 s"}.
}
\description{
A duration, in the unit that reads best
}
\keyword{internal}
