% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calib.R
\name{in_project}
\alias{in_project}
\title{A path of the configuration, seen from the working directory}
\usage{
in_project(path, root = ".")
}
\arguments{
\item{path}{a path relative to the project root.}

\item{root}{the project root.}
}
\value{
the path to use from the working directory.
}
\description{
A path of the configuration, seen from the working directory
}
\keyword{internal}
