% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkversionpackageandinstall.R
\name{install_local}
\alias{install_local}
\title{Check a tar.gz package installation and local availability of new version}
\usage{
install_local(name_package)
}
\arguments{
\item{name_package}{Character string. Name of the package to check}
}
\value{
package installation if a newer version of the tar.gz is found locally
}
\description{
Check a tar.gz package installation and local availability of new version
}
