% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_export.R
\docType{data}
\name{label_database}
\alias{label_database}
\title{Variables label database}
\format{
data.frame
}
\source{
Labels and variables excel sheet
}
\usage{
data(label_database)
}
\description{
Variables label database
}
\author{
ThreeMe team
}
\keyword{data}
