% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel_calibration.R
\name{load_excel_calibration}
\alias{load_excel_calibration}
\title{Loading a specifically formatted excel file for threeme scenarios}
\usage{
load_excel_calibration(
  excel_sheet = "configuration/scenarii_calib/scenarii_inputs.xlsx",
  sheet_to_load = "baseline",
  baseline = FALSE,
  base_year = config$input$baseyear,
  check_tol = 1e-07,
  stop_if_calib_fail = TRUE,
  keep_baseyear_calib_data = TRUE,
  calib_baseline = OGcalib,
  calib_shock = calib_new_base
)
}
\arguments{
\item{excel_sheet}{file path to the excel file}

\item{sheet_to_load}{name of the sheet in the excel file with the scenario}

\item{baseline}{Boolean. Set to true if the sheet is used to calibrate the baseline scenario. Default: FALSE}

\item{base_year}{Integer. Year of the data calibration}

\item{check_tol}{Float. Tolerance for the Verification of the match between calibration data and data in the sheet at the baseyear}

\item{stop_if_calib_fail}{Boolean. Whether to stop if there is a significant mismatch between original data and excel data. Default: TRUE}

\item{keep_baseyear_calib_data}{Boolean. If \code{stop_if_calib_fail} is false and there is a significant mismatch in data, setting to TRUE means the calibration data will be used over the excel data at the base year. Default: TRUE}

\item{calib_baseline}{Name of the calibrated data for the baseline. Default is the ermeeth2 data name "OGcalib"}

\item{calib_shock}{Name of the calibrated data after integrating the baseline scenario changes. Default is the ermeeth2 data name "calib_new_base"}
}
\value{
a data.frame containing the series from the excel sheet, without missing observation
}
\description{
The function loads series from a specified sheet in an excel file, following a determined format. The function also fills missing observations using splines
}
\examples{
\dontrun{
shock_ch <- load_excel_calibration(excel_sheet = "data/scenarii_inputs.xlsx",
sheet_to_load = "shock_NEAM",
stop_if_calib_fail = FALSE,
check_tol = 10e-8,
keep_baseyear_calib_data = TRUE)
}
}
