% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loadResults.R
\name{loadResults}
\alias{loadResults}
\title{Load ThreeME results from a csv exported by EViews
Note that the csv files must be located in the csv/ directory,
baseline results should be indexed by '_0' and scenario results by '_2',
and that all scenarios must share the same baseline}
\usage{
loadResults(
  scenarios,
  by_commodity = FALSE,
  by_sector = FALSE,
  bridge_c = NULL,
  bridge_s = NULL,
  names_s = NULL,
  names_c = NULL,
  csv_folder = "csv",
  aggregation_rules_path = file.path("bridges", "aggregation_rules.xlsx")
)
}
\arguments{
\item{scenarios}{Name of the scenario csv files (with or without a csv extension)}

\item{by_commodity}{TRUE if variables should be aggregated by commodity where applicable and given a specified bridge. (FALSE by default)}

\item{by_sector}{TRUE if variables should be aggregated by sector where applicable and given a specified bridge. (FALSE by default)}

\item{bridge_c}{bridge for commodities to be used if aggregation by commodity is requested. A bridge is a list of vectors. The names specify the larger aggregation group, and the elements of the vectors are the commodity codes that belong in this larger group.}

\item{bridge_s}{bridge for sector to be used if aggregation by sector is requested. A bridge is a list of vectors. The names specify the larger aggregation group, and the elements of the vectors are the sector codes that belong in this larger group.}

\item{names_s}{a data.frame used only if no aggregation is requested for sectors. One column "code" should contain the codes used in the sector variables and a column "name" contains the explicit name.}

\item{names_c}{a data.frame used only if no aggregation is requested for commodities. One column "code" should contain the codes used in the commodity variables and a column "name" contains the explicit name.}

\item{csv_folder}{path where to source the csv from simulations}

\item{aggregation_rules_path}{path where to find the aggregation rules table}
}
\value{
A data.frame with the following columns:
\itemize{
\item variable: name of the ThreeME variable for which the result is reported in a given row
\item year: year for which the result is reported in a given row
\item one column per scenario, named after the scenario (e.g. "baseline") containing
the values of the results themselves
\item sector : if applicable, the sector name pertaining to the variable, NA otherwise
\item commodity : if applicable, the commodity name pertaining to the variable, NA otherwise
}
}
\description{
Load ThreeME results from a csv exported by EViews
Note that the csv files must be located in the csv/ directory,
baseline results should be indexed by '_0' and scenario results by '_2',
and that all scenarios must share the same baseline
}
