% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_copy.R
\name{make_copy}
\alias{make_copy}
\title{Make a copy of a variable}
\usage{
make_copy(var_name, append = ".old", envir = .GlobalEnv, envir2 = .GlobalEnv)
}
\arguments{
\item{var_name}{character string. Object from the envir to copy as a character string}

\item{append}{character string to append to the copy. Default is ".old"}

\item{envir}{Environment from where the variable is from. Default is GlobalEnv}

\item{envir2}{Environment where to store the copy. Default is GlobalEnv.}
}
\value{
object in the environment
}
\description{
Make a copy of a variable
}
\examples{
\dontrun{
a <- 3
make_copy(a)
 }
}
