% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_parse.R
\name{mdl_parse_var}
\alias{mdl_parse_var}
\title{Parse a lone variable declaration, as found in an exovar .mdl file}
\usage{
mdl_parse_var(x)
}
\arguments{
\item{x}{character(1), e.g. \code{"ADJUST_MARKUP[s]"}.}
}
\value{
an \code{mdl_var} node.
}
\description{
Parse a lone variable declaration, as found in an exovar .mdl file
}
\examples{
mdl_parse_var("ADJUST_MARKUP[s]")$indices
}
