% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_any}
\alias{message_any}
\title{Display an informative message}
\usage{
message_any(txt_message = "TEST", custom_symbol = "")
}
\arguments{
\item{txt_message}{text to display}

\item{custom_symbol}{symbol to display before the text. The default, \code{""},
uses cli's info bullet.}
}
\value{
Called for its side effect: a message.
}
\description{
Display an informative message
}
