% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_delete}
\alias{message_delete}
\title{Display a deletion message}
\usage{
message_delete(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
Called for its side effect: a message.
}
\description{
Display a deletion message
}
