% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_main_step}
\alias{message_main_step}
\title{Display a main step heading}
\usage{
message_main_step(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
Called for its side effect: a message.
}
\description{
Display a main step heading
}
