% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_not_ok}
\alias{message_not_ok}
\title{Display a failure message}
\usage{
message_not_ok(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
Called for its side effect: a message.
}
\description{
Display a failure message
}
