% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_sub_step}
\alias{message_sub_step}
\title{Display a sub-step}
\usage{
message_sub_step(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
Called for its side effect: a message.
}
\description{
Display a sub-step
}
