% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_export.R
\docType{data}
\name{trad_database}
\alias{trad_database}
\title{##data_export
#' oilprice_fra
#'
#' full ThreeMe database without sector and commodity aggregation
#'
#' @name oilprice_fra
#' @docType data
#' @author ThreeMe team
#' @source ThreeMe simulations
#' @format data.frame
#' @keywords data
#' @usage data(oilprice_fra)
#'
#' ##Dataexport
"oilprice_fra"}
\format{
data.frame
}
\source{
Translation excel sheet
}
\usage{
data(trad_database)
}
\description{
##data_export
#' oilprice_fra_agg
#'
#' full ThreeMe database with sector and commodity aggregation
#'
#' @name oilprice_fra_agg
#' @docType data
#' @author ThreeMe team
#' @source ThreeMe simulations
#' @format data.frame
#' @keywords data
#' @usage data(oilprice_fra_agg)
#'
#' ##Dataexport
"oilprice_fra_agg"
}
\details{
Translation database for commonly found labels
}
\author{
ThreeMe team
}
\keyword{data}
