% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{trad_language_base}
\alias{trad_language_base}
\title{Database for translation with trad function}
\format{
data.frame
}
\source{
threeme team
}
\usage{
trad_language_base
}
\description{
Database for translation with trad function
}
\author{
ThreeME
}
\keyword{language}
\keyword{threeme}
