% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformation_spec}
\alias{transformation_spec}
\title{Look up one transformation specification}
\usage{
transformation_spec(transformation)
}
\arguments{
\item{transformation}{character(1) the transformation id.}
}
\value{
The matching element of \code{\link[=threeme_transformations]{threeme_transformations()}}.
}
\description{
Look up one transformation specification
}
