% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_merge.R
\name{update_data_merge}
\alias{update_data_merge}
\title{update_data_merge}
\usage{
update_data_merge(old_data, new_data, by_join_var = "year")
}
\arguments{
\item{old_data}{data.frame to update}

\item{new_data}{new data.frame holding the updated values}

\item{by_join_var}{character vector of the columns to join on. Default is "year"}
}
\value{
a data.frame with the columns of \code{old_data} updated with the values
of \code{new_data}. If the two data.frames do not have the same number of rows,
\code{old_data} is returned unchanged.
}
\description{
Combine two databases that may have variables in common. The
common variables are replaced by their values from the new database.
}
