% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_edit.R
\name{config_edit}
\alias{config_edit}
\title{Apply several edits to a configuration file}
\usage{
config_edit(file, values, out = file)
}
\arguments{
\item{file}{path to the configuration file.}

\item{values}{a named list of replacements. Each element is either R source
text (wrap it in [I()] to say "this is already code") or a value to be
deparsed.}

\item{out}{where to write. Defaults to `file`, editing in place.}
}
\value{
invisibly, the path written.
}
\description{
Apply several edits to a configuration file
}
\examples{
\dontrun{
config_edit("configuration/config_input_threeme.R",
            list(project_name = "my_run", shockyear = 2025))
}
}
