% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_parse.R
\name{mdl_dependent_name}
\alias{mdl_dependent_name}
\title{The name of an equation's dependent variable, indices included}
\usage{
mdl_dependent_name(eq)
}
\arguments{
\item{eq}{an `mdl_eq` node, from [mdl_parse()].}
}
\value{
character(1), e.g. `"F_n[f, s]"`, or `NA_character_`.
}
\description{
The name of an equation's dependent variable, indices included
}
\examples{
mdl_dependent_name(mdl_parse("F[K, s] = (1 - Rdep[s]) * F[K, s]{-1}"))
}
