% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load.R
\name{readconfig}
\alias{readconfig}
\title{Read the configuration files and create the config list}
\usage{
readconfig(
  input_config_file = file.path("configuration", "config_input_threeme.R"),
  output_config_file = file.path("configuration", "config_output_threeme.R")
)
}
\arguments{
\item{input_config_file}{file path to the configuration input file}

\item{output_config_file}{file path to the configuration output file}
}
\value{
a config list with all the necessary elements to run the simulations
}
\description{
Read the configuration files and create the config list
}
\examples{
\dontrun{
config <- readconfig <- function(
input_config_file = file.path("configuration", "config_input_threeme.R") ,
output_config_file = file.path("configuration", "config_output_threeme.R") )
}

}
